[core] Add data_type definitions to properties that reference them
Puppet doesn't sync the data types to the agent. Because we use the data from the data type to determine if the property is a reference property, things don't work when used with a puppet agent.

I have investigated if I can force usage of the data types, but that doesn't seem to work.

The work around now is to use the FUL definition without the typdef in the properties. We do this by changing the property generators